spacesuit: move SignedInteger into a separate module
